chore: replace_versionを別ファイルに分割する
