+++ .github/workflows/build_and_release_infer.yml
@@ -54,7 +54,8 @@ jobs:
       - name: Build e2k-py
         working-directory: infer/tools
         run: |
-          uv run build_e2k_py.py --version ${{ inputs.version }} ${{ matrix.build_args }}
+          uv run replace_version.py --version ${{ inputs.version }}
+          uv run build_e2k_py.py ${{ matrix.build_args }}
 
       - name: Upload artifacts
         uses: actions/upload-artifact@v4
