+++ infer/tools/paths.py
@@ -1,0 +1,3 @@
+from pathlib import Path
+
+infer_root = Path(__file__).parent.parent
